Added powerline font configuration.
Note that powerline fonts need to be installed separately from this
commit; see notes in README.md on how to get them.
Added line to `.xinitrc` to look in the user home directory for fonts.
Added line to nvim init.vim to set encoding to utf-8 in order to take
advantage of the powerline fonts.
Changed urxvt font in `.Xresources` to source code pro for powerline.